+++ skued/simulation/__init__.py
@@ -3,3 +3,4 @@
 from .potential import electrostatic, pelectrostatic
 from .powdersim import powdersim
 from .structure_factors import affe, structure_factor
+from .kinematic import kinematicsim