+++ CHANGELOG.rst
@@ -4,6 +4,8 @@ Changelog
 Release 2.0.5
 -------------
 
+* Added `skued.kinematicsim`, a simple function to compute electron diffraction patterns from 
+  crystals structures in the kinematic approximation (i.e. thin samples).
 * Various documentation improvements and fixes.
 
 Release 2.0.4